refactor type enum name
