+++ src/__tests__/axios/api.test.ts
@@ -2,7 +2,7 @@ import { AxiosError } from "axios";
 import MockAdapter from "axios-mock-adapter";
 
 import api, { API_BASE_URL } from "@/axios";
-import { HttpResponse } from "@/types/enum/HttpResponse";
+import { HttpResponse } from "@/types/enum/http-response";
 import { showToast } from "@/utils/toast";
 
 jest.mock("@/utils/toast");
